requirements.txt
# ==============================================================================
# CORE DJANGO - Essentiels pour GreenCart MVP
# ==============================================================================
Django==5.2.4
python-decouple==3.8
dj-database-url==3.0.1

# ==============================================================================
# API & REST FRAMEWORK - Pour l'API GreenCart
# ==============================================================================
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.7.0
drf-spectacular==0.28.0

# ==============================================================================
# IMAGES & MEDIA - Pour les photos des produits
# ==============================================================================
Pillow==11.3.0

# ==============================================================================
# API DOCUMENTATION - Pour Swagger/OpenAPI
# ==============================================================================
drf-spectacular==0.27.0